wifi: Fix crash/double free warning in TX
Without taking a reference driver is just putting down the reference
once the packet is copied to a local buffer, depending on the timing
of the networking stack timing to actually free this buffer it can lead
to crash or double freeing warning (when packet debugging is enabled).

This is readily evident with CONFIG_NET_TC_TX_COUNT=1 where the packet
processing is done in a thread, so, the delay will 100% cause the crash
as immediately after l2_send, the packet buffer is accessed.

There is no need to drop the reference, as driver has not taken one, so,
packet will be freed automatically by the networking stack.