wifi: Fix race condition in key installation
With the introduction of queue in the networking stack M4 can be delayed
and if the key is installed before M4 is processed, then RPU is
encrypting M4 with new key but as the new key is not yet installed at
the AP, M4 is droppped.

This is a classic M4 installation race, that exists even without this
change, but adding delay to M4 worsens it. So, for now we can disable
the queue, and address this later.

This was bisected to commit 143745bad ("Tweak system settings for performance").

Signed-off-by: Krishna T <krishna.t@nordicsemi.no>