drivers: wifi: Fix sleep checks
For non-sleep the RD/WR status register access is mainly for
triggerring/waiting, so, all return values are hardcoded to zero, but
for sleep the code relies on the return value read from the RDSR value,
so, this causes a false negative wakeup failure.

Fix the return values for public APIs to return the actual register
being accessed.

Signed-off-by: Krishna T <krishna.t@nordicsemi.no>