drivers: wifi: Add a workaround for some EK boards
In some EK boards we are seeing that the RPU powering on takes a bit
more time and if we start waking up RPU from its initial sleep, then the
LMAC/UMAC boot failure was observed.

As a temporary workaround add an extra wait time after powering on but
before waking up, after a few trails 8ms is the most reliable wait time.

Signed-off-by: Krishna T <krishna.t@nordicsemi.no>